modify version 1.0.6

+++ YWaterfallFlowView/JZWaterfallFlowViewController.m
@@ -117,7 +117,7 @@
     [_dicEachColumnMaxBottom[@"0"] floatValue];
     NSUInteger MinColumn = 0;
     for (NSUInteger i = 1; i < _cloumnNumber; i ++) {
-        NSString *key = [NSString stringWithFormat:@"%lu",i];
+        NSString *key = [NSString stringWithFormat:@"%lu", (unsigned long)i];
         if ([_dicEachColumnMaxBottom[key] floatValue] < MinOfsset)
         {
             MinColumn = i;
@@ -132,7 +132,7 @@
     CGFloat MaxOfsset =
     [_dicEachColumnMaxBottom[@"0"] floatValue];
     for (NSUInteger i = 1; i < _cloumnNumber; i ++) {
-        NSString *key = [NSString stringWithFormat:@"%lu",i];
+        NSString *key = [NSString stringWithFormat:@"%lu", (unsigned long)i];
         if ([_dicEachColumnMaxBottom[key] floatValue] > MaxOfsset)
         {
             MaxOfsset = [_dicEachColumnMaxBottom[key] floatValue];
@@ -266,7 +266,7 @@
                 cell = [self.WDataSource waterfallFlow:self cellForRowAtIndex:i];
                 // set cell's frame.
                 NSUInteger currentColumn            = [self getCurrentColumn];
-                NSString *curentColumnKey           = [NSString stringWithFormat:@"%lu",currentColumn];
+                NSString *curentColumnKey           = [NSString stringWithFormat:@"%lu", (unsigned long)currentColumn];
                 CGFloat currentColumnLastMaxOffsetY = [_dicEachColumnMaxBottom[curentColumnKey] floatValue];
                 
                 if (_dicEachCellY[key]) {
